Move the query module into db subpackage
